+++ Sources/AutoSDK/AutoSDKVersion.m
@@ -1,4 +1,4 @@
 #import "include/AutoSDK.h"
 
-double AutoSDKVersionNumber = 1.5;
-const unsigned char AutoSDKVersionString[] = "1.5.0";
+double AutoSDKVersionNumber = 1.6;
+const unsigned char AutoSDKVersionString[] = "1.6.0";